Added int casting to size_t longs to silence warnings that it was losing the precision of the longs

+++ Game.c
@@ -16,7 +16,7 @@ void disposeGame(Game* game) {
 }
 
 int getTurnNumber(Game* game) {
-    return game->currentTurn;
+    return (int)game->currentTurn;
 }
 
 PlayerId getWhoseTurn(Game* game) {
